[js/common] refactor tensor in onnxruntime-common
