+++ js/common/lib/tensor-impl.ts
@@ -1,7 +1,12 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-import {Tensor as TensorInterface, TensorFromImageOptions, TensorToImageDataOptions} from './tensor';
+import {Tensor as TensorInterface} from './tensor';
+import {TensorToDataUrlOptions, TensorToImageDataOptions} from './tensor-conversion';
+import {tensorToDataURL, tensorToImageData} from './tensor-conversion-impl';
+import {TensorFromImageBitmapOptions, TensorFromImageDataOptions, TensorFromImageElementOptions, TensorFromUrlOptions} from './tensor-factory';
+import {tensorFromImage} from './tensor-factory-impl';
+import {calculateSize, tensorReshape} from './tensor-utils-impl';
 
 type TensorType = TensorInterface.Type;
 type TensorDataType = TensorInterface.DataType;
@@ -58,25 +63,6 @@ const checkBigInt = () => {
   }
 };
 
-/**
- * calculate size from dims.
- *
- * @param dims the dims array. May be an illegal input.
- */
-const calculateSize = (dims: readonly unknown[]): number => {
-  let size = 1;
-  for (let i = 0; i < dims.length; i++) {
-    const dim = dims[i];
-    if (typeof dim !== 'number' || !Number.isSafeInteger(dim)) {
-      throw new TypeError(`dims[${i}] must be an integer, got: ${dim}`);
-    }
-    if (dim < 0) {
-      throw new RangeError(`dims[${i}] must be a non-negative integer, got: ${dim}`);
-    }
-    size *= dim;
-  }
-  return size;
-};
 
 export class Tensor implements TensorInterface {
   // #region constructors
@@ -178,469 +164,30 @@ export class Tensor implements TensorInterface {
     this.size = size;
   }
   // #endregion
-  /**
-   * Create a new tensor object from image object
-   *
-   * @param buffer - Extracted image buffer data - assuming RGBA format
-   * @param imageFormat - input image configuration - required configurations height, width, format
-   * @param tensorFormat - output tensor configuration - Default is RGB format
-   */
-  private static bufferToTensor(buffer: Uint8ClampedArray|undefined, options: TensorFromImageOptions): Tensor {
-    if (buffer === undefined) {
-      throw new Error('Image buffer must be defined');
-    }
-    if (options.height === undefined || options.width === undefined) {
-      throw new Error('Image height and width must be defined');
-    }
-    if (options.tensorLayout === 'NHWC') {
-      throw new Error('NHWC Tensor layout is not supported yet');
-    }
-
-    const {height, width} = options;
-
-    const norm = options.norm ?? {mean: 255, bias: 0};
-    let normMean: [number, number, number, number];
-    let normBias: [number, number, number, number];
-
-    if (typeof (norm.mean) === 'number') {
-      normMean = [norm.mean, norm.mean, norm.mean, norm.mean];
-    } else {
-      normMean = [norm.mean![0], norm.mean![1], norm.mean![2], norm.mean![3] ?? 255];
-    }
-
-    if (typeof (norm.bias) === 'number') {
-      normBias = [norm.bias, norm.bias, norm.bias, norm.bias];
-    } else {
-      normBias = [norm.bias![0], norm.bias![1], norm.bias![2], norm.bias![3] ?? 0];
-    }
-
-    const inputformat = options.bitmapFormat !== undefined ? options.bitmapFormat : 'RGBA';
-    // default value is RGBA since imagedata and HTMLImageElement uses it
-
-    const outputformat = options.tensorFormat !== undefined ?
-        (options.tensorFormat !== undefined ? options.tensorFormat : 'RGB') :
-        'RGB';
-    const stride = height * width;
-    const float32Data = outputformat === 'RGBA' ? new Float32Array(stride * 4) : new Float32Array(stride * 3);
-
-    // Default pointer assignments
-    let step = 4, rImagePointer = 0, gImagePointer = 1, bImagePointer = 2, aImagePointer = 3;
-    let rTensorPointer = 0, gTensorPointer = stride, bTensorPointer = stride * 2, aTensorPointer = -1;
-
-    // Updating the pointer assignments based on the input image format
-    if (inputformat === 'RGB') {
-      step = 3;
-      rImagePointer = 0;
-      gImagePointer = 1;
-      bImagePointer = 2;
-      aImagePointer = -1;
-    }
-
-    // Updating the pointer assignments based on the output tensor format
-    if (outputformat === 'RGBA') {
-      aTensorPointer = stride * 3;
-    } else if (outputformat === 'RBG') {
-      rTensorPointer = 0;
-      bTensorPointer = stride;
-      gTensorPointer = stride * 2;
-    } else if (outputformat === 'BGR') {
-      bTensorPointer = 0;
-      gTensorPointer = stride;
-      rTensorPointer = stride * 2;
-    }
-
-    for (let i = 0; i < stride;
-         i++, rImagePointer += step, bImagePointer += step, gImagePointer += step, aImagePointer += step) {
-      float32Data[rTensorPointer++] = (buffer[rImagePointer] + normBias[0]) / normMean[0];
-      float32Data[gTensorPointer++] = (buffer[gImagePointer] + normBias[1]) / normMean[1];
-      float32Data[bTensorPointer++] = (buffer[bImagePointer] + normBias[2]) / normMean[2];
-      if (aTensorPointer !== -1 && aImagePointer !== -1) {
-        float32Data[aTensorPointer++] = (buffer[aImagePointer] + normBias[3]) / normMean[3];
-      }
-    }
-
-    // Float32Array -> ort.Tensor
-    const outputTensor = outputformat === 'RGBA' ? new Tensor('float32', float32Data, [1, 4, height, width]) :
-                                                   new Tensor('float32', float32Data, [1, 3, height, width]);
-    return outputTensor;
-  }
 
   // #region factory
-  static async fromImage(imageData: ImageData, options?: TensorFromImageOptions): Promise<Tensor>;
-  static async fromImage(imageElement: HTMLImageElement, options?: TensorFromImageOptions): Promise<Tensor>;
-  static async fromImage(bitmap: ImageBitmap, options: TensorFromImageOptions): Promise<Tensor>;
-  static async fromImage(urlSource: string, options?: TensorFromImageOptions): Promise<Tensor>;
+  static async fromImage(imageData: ImageData, options?: TensorFromImageDataOptions): Promise<Tensor>;
+  static async fromImage(imageElement: HTMLImageElement, options?: TensorFromImageElementOptions): Promise<Tensor>;
+  static async fromImage(bitmap: ImageBitmap, options: TensorFromImageBitmapOptions): Promise<Tensor>;
+  static async fromImage(urlSource: string, options?: TensorFromUrlOptions): Promise<Tensor>;
 
-  static async fromImage(image: ImageData|HTMLImageElement|ImageBitmap|string, options?: TensorFromImageOptions):
-      Promise<Tensor> {
-    // checking the type of image object
-    const isHTMLImageEle = typeof (HTMLImageElement) !== 'undefined' && image instanceof HTMLImageElement;
-    const isImageDataEle = typeof (ImageData) !== 'undefined' && image instanceof ImageData;
-    const isImageBitmap = typeof (ImageBitmap) !== 'undefined' && image instanceof ImageBitmap;
-    const isString = typeof image === 'string';
-
-    let data: Uint8ClampedArray|undefined;
-    let tensorConfig: TensorFromImageOptions = options ?? {};
-
-    // filling and checking image configuration options
-    if (isHTMLImageEle) {
-      // HTMLImageElement - image object - format is RGBA by default
-      const canvas = document.createElement('canvas');
-      canvas.width = image.width;
-      canvas.height = image.height;
-      const pixels2DContext = canvas.getContext('2d');
-
-      if (pixels2DContext != null) {
-        let height = image.height;
-        let width = image.width;
-        if (options !== undefined && options.resizedHeight !== undefined && options.resizedWidth !== undefined) {
-          height = options.resizedHeight;
-          width = options.resizedWidth;
-        }
-
-        if (options !== undefined) {
-          tensorConfig = options;
-          if (options.tensorFormat !== undefined) {
-            throw new Error('Image input config format must be RGBA for HTMLImageElement');
-          } else {
-            tensorConfig.tensorFormat = 'RGBA';
-          }
-          if (options.height !== undefined && options.height !== height) {
-            throw new Error('Image input config height doesn\'t match HTMLImageElement height');
-          } else {
-            tensorConfig.height = height;
-          }
-          if (options.width !== undefined && options.width !== width) {
-            throw new Error('Image input config width doesn\'t match HTMLImageElement width');
-          } else {
-            tensorConfig.width = width;
-          }
-        } else {
-          tensorConfig.tensorFormat = 'RGBA';
-          tensorConfig.height = height;
-          tensorConfig.width = width;
-        }
-
-        pixels2DContext.drawImage(image, 0, 0);
-        data = pixels2DContext.getImageData(0, 0, width, height).data;
-      } else {
-        throw new Error('Can not access image data');
-      }
-
-    } else if (isImageDataEle) {
-      // ImageData - image object - format is RGBA by default
-      const format = 'RGBA';
-      let height: number;
-      let width: number;
-
-      if (options !== undefined && options.resizedWidth !== undefined && options.resizedHeight !== undefined) {
-        height = options.resizedHeight;
-        width = options.resizedWidth;
-      } else {
-        height = image.height;
-        width = image.width;
-      }
-
-      if (options !== undefined) {
-        tensorConfig = options;
-        if (options.bitmapFormat !== undefined && options.bitmapFormat !== format) {
-          throw new Error('Image input config format must be RGBA for ImageData');
-        } else {
-          tensorConfig.bitmapFormat = 'RGBA';
-        }
-      } else {
-        tensorConfig.bitmapFormat = 'RGBA';
-      }
-
-      tensorConfig.height = height;
-      tensorConfig.width = width;
-
-      if (options !== undefined) {
-        const tempCanvas = document.createElement('canvas');
-
-        tempCanvas.width = width;
-        tempCanvas.height = height;
-
-        const pixels2DContext = tempCanvas.getContext('2d');
-
-        if (pixels2DContext != null) {
-          pixels2DContext.putImageData(image, 0, 0);
-          data = pixels2DContext.getImageData(0, 0, width, height).data;
-        } else {
-          throw new Error('Can not access image data');
-        }
-      } else {
-        data = image.data;
-      }
-
-    } else if (isImageBitmap) {
-      // ImageBitmap - image object - format must be provided by user
-      if (options === undefined) {
-        throw new Error('Please provide image config with format for Imagebitmap');
-      }
-      if (options.bitmapFormat !== undefined) {
-        throw new Error('Image input config format must be defined for ImageBitmap');
-      }
-
-      const pixels2DContext = document.createElement('canvas').getContext('2d');
-
-      if (pixels2DContext != null) {
-        const height = image.height;
-        const width = image.width;
-        pixels2DContext.drawImage(image, 0, 0, width, height);
-        data = pixels2DContext.getImageData(0, 0, width, height).data;
-        if (options !== undefined) {
-          // using square brackets to avoid TS error - type 'never'
-          if (options.height !== undefined && options.height !== height) {
-            throw new Error('Image input config height doesn\'t match ImageBitmap height');
-          } else {
-            tensorConfig.height = height;
-          }
-          // using square brackets to avoid TS error - type 'never'
-          if (options.width !== undefined && options.width !== width) {
-            throw new Error('Image input config width doesn\'t match ImageBitmap width');
-          } else {
-            tensorConfig.width = width;
-          }
-        } else {
-          tensorConfig.height = height;
-          tensorConfig.width = width;
-        }
-        return Tensor.bufferToTensor(data, tensorConfig);
-      } else {
-        throw new Error('Can not access image data');
-      }
-
-    } else if (isString) {
-      return new Promise((resolve, reject) => {
-        const canvas = document.createElement('canvas');
-        const context = canvas.getContext('2d');
-        if (!image || !context) {
-          return reject();
-        }
-        const newImage = new Image();
-        newImage.crossOrigin = 'Anonymous';
-        newImage.src = image;
-        newImage.onload = () => {
-          canvas.width = newImage.width;
-          canvas.height = newImage.height;
-          context.drawImage(newImage, 0, 0, canvas.width, canvas.height);
-          const img = context.getImageData(0, 0, canvas.width, canvas.height);
-          if (options !== undefined) {
-            if (options.height !== undefined && options.height !== canvas.height) {
-              throw new Error('Image input config height doesn\'t match height');
-            } else {
-              tensorConfig.height = canvas.height;
-            }
-            if (options.width !== undefined && options.width !== canvas.width) {
-              throw new Error('Image input config width doesn\'t match width');
-            } else {
-              tensorConfig.width = canvas.width;
-            }
-          } else {
-            tensorConfig.height = canvas.height;
-            tensorConfig.width = canvas.width;
-          }
-          resolve(Tensor.bufferToTensor(img.data, tensorConfig));
-        };
-      });
-    } else {
-      throw new Error('Input data provided is not supported - aborted tensor creation');
-    }
-
-    if (data !== undefined) {
-      return Tensor.bufferToTensor(data, tensorConfig);
-    } else {
-      throw new Error('Input data provided is not supported - aborted tensor creation');
-    }
+  static async fromImage(
+      image: ImageData|HTMLImageElement|ImageBitmap|string,
+      options?: TensorFromImageDataOptions|TensorFromImageElementOptions|TensorFromImageBitmapOptions|
+      TensorFromUrlOptions): Promise<Tensor> {
+    return tensorFromImage(image, options);
   }
+  // #endregion
 
-  toDataURL(options?: TensorToImageDataOptions): string {
-    const canvas = document.createElement('canvas');
-    canvas.width = this.dims[3];
-    canvas.height = this.dims[2];
-    const pixels2DContext = canvas.getContext('2d');
-
-    if (pixels2DContext != null) {
-      // Default values for height and width & format
-      let width: number;
-      let height: number;
-      if (options?.tensorLayout !== undefined && options.tensorLayout === 'NHWC') {
-        width = this.dims[2];
-        height = this.dims[3];
-      } else {  // Default layout is NCWH
-        width = this.dims[3];
-        height = this.dims[2];
-      }
-
-      const inputformat = options?.format !== undefined ? options.format : 'RGB';
-
-      const norm = options?.norm;
-      let normMean: [number, number, number, number];
-      let normBias: [number, number, number, number];
-      if (norm === undefined || norm.mean === undefined) {
-        normMean = [255, 255, 255, 255];
-      } else {
-        if (typeof (norm.mean) === 'number') {
-          normMean = [norm.mean, norm.mean, norm.mean, norm.mean];
-        } else {
-          normMean = [norm.mean[0], norm.mean[1], norm.mean[2], 0];
-          if (norm.mean[3] !== undefined) {
-            normMean[3] = norm.mean[3];
-          }
-        }
-      }
-      if (norm === undefined || norm.bias === undefined) {
-        normBias = [0, 0, 0, 0];
-      } else {
-        if (typeof (norm.bias) === 'number') {
-          normBias = [norm.bias, norm.bias, norm.bias, norm.bias];
-        } else {
-          normBias = [norm.bias[0], norm.bias[1], norm.bias[2], 0];
-          if (norm.bias[3] !== undefined) {
-            normBias[3] = norm.bias[3];
-          }
-        }
-      }
-
-      const stride = height * width;
-      // Default pointer assignments
-      let rTensorPointer = 0, gTensorPointer = stride, bTensorPointer = stride * 2, aTensorPointer = -1;
-
-      // Updating the pointer assignments based on the input image format
-      if (inputformat === 'RGBA') {
-        rTensorPointer = 0;
-        gTensorPointer = stride;
-        bTensorPointer = stride * 2;
-        aTensorPointer = stride * 3;
-      } else if (inputformat === 'RGB') {
-        rTensorPointer = 0;
-        gTensorPointer = stride;
-        bTensorPointer = stride * 2;
-      } else if (inputformat === 'RBG') {
-        rTensorPointer = 0;
-        bTensorPointer = stride;
-        gTensorPointer = stride * 2;
-      }
-
-      for (let i = 0; i < height; i++) {
-        for (let j = 0; j < width; j++) {
-          const R = ((this.data[rTensorPointer++] as number) - normBias[0]) * normMean[0];  // R value
-          const G = ((this.data[gTensorPointer++] as number) - normBias[1]) * normMean[1];  // G value
-          const B = ((this.data[bTensorPointer++] as number) - normBias[2]) * normMean[2];  // B value
-          const A = aTensorPointer === -1 ?
-              255 :
-              ((this.data[aTensorPointer++] as number) - normBias[3]) * normMean[3];  // A value
-          // eslint-disable-next-line @typescript-eslint/restrict-plus-operands
-          pixels2DContext.fillStyle = 'rgba(' + R + ',' + G + ',' + B + ',' + A + ')';
-          pixels2DContext.fillRect(j, i, 1, 1);
-        }
-      }
-      return canvas.toDataURL();
-    } else {
-      throw new Error('Can not access image data');
-    }
+  // #region conversions
+  toDataURL(options?: TensorToDataUrlOptions): string {
+    return tensorToDataURL(this, options);
   }
 
   toImageData(options?: TensorToImageDataOptions): ImageData {
-    const pixels2DContext = document.createElement('canvas').getContext('2d');
-    let image: ImageData;
-    if (pixels2DContext != null) {
-      // Default values for height and width & format
-      let width: number;
-      let height: number;
-      let channels: number;
-      if (options?.tensorLayout !== undefined && options.tensorLayout === 'NHWC') {
-        width = this.dims[2];
-        height = this.dims[1];
-        channels = this.dims[3];
-      } else {  // Default layout is NCWH
-        width = this.dims[3];
-        height = this.dims[2];
-        channels = this.dims[1];
-      }
-      const inputformat = options !== undefined ? (options.format !== undefined ? options.format : 'RGB') : 'RGB';
-
-      const norm = options?.norm;
-      let normMean: [number, number, number, number];
-      let normBias: [number, number, number, number];
-      if (norm === undefined || norm.mean === undefined) {
-        normMean = [255, 255, 255, 255];
-      } else {
-        if (typeof (norm.mean) === 'number') {
-          normMean = [norm.mean, norm.mean, norm.mean, norm.mean];
-        } else {
-          normMean = [norm.mean[0], norm.mean[1], norm.mean[2], 255];
-          if (norm.mean[3] !== undefined) {
-            normMean[3] = norm.mean[3];
-          }
-        }
-      }
-      if (norm === undefined || norm.bias === undefined) {
-        normBias = [0, 0, 0, 0];
-      } else {
-        if (typeof (norm.bias) === 'number') {
-          normBias = [norm.bias, norm.bias, norm.bias, norm.bias];
-        } else {
-          normBias = [norm.bias[0], norm.bias[1], norm.bias[2], 0];
-          if (norm.bias[3] !== undefined) {
-            normBias[3] = norm.bias[3];
-          }
-        }
-      }
-
-      const stride = height * width;
-      if (options !== undefined) {
-        if (options.height !== undefined && options.height !== height) {
-          throw new Error('Image output config height doesn\'t match tensor height');
-        }
-        if (options.width !== undefined && options.width !== width) {
-          throw new Error('Image output config width doesn\'t match tensor width');
-        }
-        if (options.format !== undefined && (channels === 4 && options.format !== 'RGBA') ||
-            (channels === 3 && (options.format !== 'RGB' && options.format !== 'BGR'))) {
-          throw new Error('Tensor format doesn\'t match input tensor dims');
-        }
-      }
-
-      // Default pointer assignments
-      const step = 4;
-      let rImagePointer = 0, gImagePointer = 1, bImagePointer = 2, aImagePointer = 3;
-      let rTensorPointer = 0, gTensorPointer = stride, bTensorPointer = stride * 2, aTensorPointer = -1;
-
-      // Updating the pointer assignments based on the input image format
-      if (inputformat === 'RGBA') {
-        rTensorPointer = 0;
-        gTensorPointer = stride;
-        bTensorPointer = stride * 2;
-        aTensorPointer = stride * 3;
-      } else if (inputformat === 'RGB') {
-        rTensorPointer = 0;
-        gTensorPointer = stride;
-        bTensorPointer = stride * 2;
-      } else if (inputformat === 'RBG') {
-        rTensorPointer = 0;
-        bTensorPointer = stride;
-        gTensorPointer = stride * 2;
-      }
-
-      image = pixels2DContext.createImageData(width, height);
-
-      for (let i = 0; i < height * width;
-           rImagePointer += step, gImagePointer += step, bImagePointer += step, aImagePointer += step, i++) {
-        image.data[rImagePointer] = ((this.data[rTensorPointer++] as number) - normBias[0]) * normMean[0];  // R value
-        image.data[gImagePointer] = ((this.data[gTensorPointer++] as number) - normBias[1]) * normMean[1];  // G value
-        image.data[bImagePointer] = ((this.data[bTensorPointer++] as number) - normBias[2]) * normMean[2];  // B value
-        image.data[aImagePointer] = aTensorPointer === -1 ?
-            255 :
-            ((this.data[aTensorPointer++] as number) - normBias[3]) * normMean[3];  // A value
-      }
-
-    } else {
-      throw new Error('Can not access image data');
-    }
-    return image;
+    return tensorToImageData(this, options);
   }
+  // #endregion
 
   // #region fields
   readonly dims: readonly number[];
@@ -651,7 +198,7 @@ export class Tensor implements TensorInterface {
 
   // #region tensor utilities
   reshape(dims: readonly number[]): Tensor {
-    return new Tensor(this.type, this.data, dims);
+    return tensorReshape(this, dims);
   }
   // #endregion
 }
